usb: dwc3-msm: Don't issue power supply change for invalid charger
This is constantly causing wakeups and general havoc/instability, as it
constantly issues a power supply change for the same power supply type
(DWC3_INVALID_CHARGER).

Constantly issuing a power supply change for an invalid charger doesn't
make sense, so don't do it in order to fix the USB madness.

Change-Id: I4e3688b52abde094a421d661d34e7cff3058e577
Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: Vaisakh Murali <vaisakhmurali@gmail.com>